+++ src/main/java/fll/web/schedule/AddScheduleTeams.java
@@ -1,6 +1,6 @@
 /*
- * Copyright (c) 2010 INSciTE.  All rights reserved
- * INSciTE is on the web at: http://www.hightechkids.org
+ * Copyright (c) 2019 High Tech Kids.  All rights reserved
+ * HighTechKids is on the web at: http://www.hightechkids.org
  * This code is released under GPL; see LICENSE.txt for details.
  */
 
@@ -9,7 +9,7 @@ package fll.web.schedule;
 import java.io.IOException;
 import java.sql.Connection;
 import java.sql.SQLException;
-import java.util.List;
+import java.util.Collection;
 
 import javax.servlet.ServletContext;
 import javax.servlet.ServletException;
@@ -19,22 +19,19 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import javax.sql.DataSource;
 
-
-
 import fll.db.Queries;
-import fll.util.FLLRuntimeException;
-
+import fll.scheduler.TeamScheduleInfo;
+import fll.util.FLLInternalException;
 import fll.web.ApplicationAttributes;
 import fll.web.BaseFLLServlet;
 import fll.web.SessionAttributes;
 import fll.web.WebUtils;
-import net.mtu.eggplant.util.sql.SQLFunctions;
 
 /**
- * Commit the event divisions for the current tournament.
+ * Add teams to the database that are in the schedule.
  */
-@WebServlet("/schedule/CommitEventDivisions")
-public class CommitEventDivisions extends BaseFLLServlet {
+@WebServlet("/schedule/AddScheduleTeams")
+public class AddScheduleTeams extends BaseFLLServlet {
 
   private static final org.apache.logging.log4j.Logger LOGGER = org.apache.logging.log4j.LogManager.getLogger();
 
@@ -44,29 +41,34 @@ public class CommitEventDivisions extends BaseFLLServlet {
                                 final ServletContext application,
                                 final HttpSession session)
       throws IOException, ServletException {
-    final DataSource datasource = ApplicationAttributes.getDataSource(application);
     final UploadScheduleData uploadScheduleData = SessionAttributes.getNonNullAttribute(session, UploadScheduleData.KEY,
                                                                                         UploadScheduleData.class);
 
-    Connection connection = null;
-    try {
-      connection = datasource.getConnection();
+    final DataSource datasource = ApplicationAttributes.getDataSource(application);
+    try (Connection connection = datasource.getConnection()) {
       final int tournamentID = Queries.getCurrentTournament(connection);
 
-      final List<EventDivisionInfo> eventDivisionInfo = uploadScheduleData.getEventDivisionInfo();
+      final Collection<TeamScheduleInfo> missingTeams = uploadScheduleData.getMissingTeams();
+      missingTeams.stream().forEach(ti -> {
+        try {
+          Queries.addTeam(connection, ti.getTeamNumber(), ti.getTeamName(), ti.getOrganization());
 
-      for (final EventDivisionInfo info : eventDivisionInfo) {
-        Queries.updateTeamEventDivision(connection, info.getTeamNumber(), tournamentID, info.getEventDivision());
-      }
+          Queries.addTeamToTournament(connection, ti.getTeamNumber(), tournamentID, ti.getAwardGroup(),
+                                      ti.getJudgingGroup());
+        } catch (final SQLException e) {
+          final String message = "Error talking to the database";
+          LOGGER.error(message, e);
+          throw new FLLInternalException(message, e);
+        }
+      });
 
-      WebUtils.sendRedirect(application, response, "/schedule/CommitSchedule");
-      return;
+      WebUtils.sendRedirect(application, response, "/schedule/CheckMissingTeams");
+
     } catch (final SQLException e) {
-      LOGGER.error("There was an error talking to the database", e);
-      throw new FLLRuntimeException("There was an error talking to the database", e);
-    } finally {
-      SQLFunctions.close(connection);
+      final String message = "Error talking to the database";
+      LOGGER.error(message, e);
+      throw new FLLInternalException(message, e);
     }
-
   }
+
 }