+++ src/main/java/fll/web/schedule/CheckMissingTeams.java
@@ -1,6 +1,6 @@
 /*
- * Copyright (c) 2010 INSciTE.  All rights reserved
- * INSciTE is on the web at: http://www.hightechkids.org
+ * Copyright (c) 2019 High Tech Kids.  All rights reserved
+ * HighTechKids is on the web at: http://www.hightechkids.org
  * This code is released under GPL; see LICENSE.txt for details.
  */
 
@@ -10,7 +10,6 @@ import java.io.IOException;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.Collection;
-import java.util.List;
 
 import javax.servlet.ServletContext;
 import javax.servlet.ServletException;
@@ -21,24 +20,19 @@ import javax.servlet.http.HttpSession;
 import javax.sql.DataSource;
 
 import fll.db.Queries;
-import fll.scheduler.ConstraintViolation;
-import fll.scheduler.SchedParams;
-import fll.scheduler.ScheduleChecker;
-import fll.scheduler.SubjectiveStation;
+import fll.scheduler.TeamScheduleInfo;
 import fll.scheduler.TournamentSchedule;
-import fll.util.FLLRuntimeException;
+import fll.util.FLLInternalException;
 import fll.web.ApplicationAttributes;
 import fll.web.BaseFLLServlet;
 import fll.web.SessionAttributes;
 import fll.web.WebUtils;
 
 /**
- * Read the uploaded file, then check for constraint
- * violations. Stores the schedule in
- * {@link UploadScheduleData#setSchedule(TournamentSchedule)}.
+ * Check for teams that are in the schedule and not in the database.
  */
-@WebServlet("/schedule/CheckViolations")
-public class CheckViolations extends BaseFLLServlet {
+@WebServlet("/schedule/CheckMissingTeams")
+public class CheckMissingTeams extends BaseFLLServlet {
 
   private static final org.apache.logging.log4j.Logger LOGGER = org.apache.logging.log4j.LogManager.getLogger();
 
@@ -57,37 +51,23 @@ public class CheckViolations extends BaseFLLServlet {
     final DataSource datasource = ApplicationAttributes.getDataSource(application);
     try (Connection connection = datasource.getConnection()) {
 
-      final List<SubjectiveStation> subjectiveStations = uploadScheduleData.getSubjectiveStations();
-
       final int tournamentID = Queries.getCurrentTournament(connection);
-      final Collection<ConstraintViolation> violations = schedule.compareWithDatabase(connection, tournamentID);
-      final SchedParams schedParams = new SchedParams(subjectiveStations, SchedParams.DEFAULT_PERFORMANCE_MINUTES,
-                                                      SchedParams.MINIMUM_CHANGETIME_MINUTES,
-                                                      SchedParams.MINIMUM_PERFORMANCE_CHANGETIME_MINUTES);
-      final ScheduleChecker checker = new ScheduleChecker(schedParams, schedule);
-      violations.addAll(checker.verifySchedule());
-      if (violations.isEmpty()) {
-        WebUtils.sendRedirect(application, response, "/schedule/GatherTeamInformationChanges");
-        return;
-      } else {
-        uploadScheduleData.setViolations(violations);
-        for (final ConstraintViolation violation : violations) {
-          if (ConstraintViolation.Type.HARD == violation.getType()) {
-            WebUtils.sendRedirect(application, response, "/schedule/displayHardViolations.jsp");
-            return;
-          }
-        }
-        WebUtils.sendRedirect(application, response, "/schedule/displaySoftViolations.jsp");
-        return;
-      }
 
+      final Collection<TeamScheduleInfo> missingTeams = schedule.findTeamsNotInDatabase(connection, tournamentID);
+      if (!missingTeams.isEmpty()) {
+        uploadScheduleData.setMissingTeams(missingTeams);
+        WebUtils.sendRedirect(application, response, "/schedule/promptAddScheduleTeams.jsp");
+      } else {
+        WebUtils.sendRedirect(application, response, "/schedule/CheckViolations");
+      }
     } catch (final SQLException e) {
       final String message = "Error talking to the database";
       LOGGER.error(message, e);
-      throw new FLLRuntimeException(message, e);
+      throw new FLLInternalException(message, e);
     } finally {
       session.setAttribute(UploadScheduleData.KEY, uploadScheduleData);
     }
+
   }
 
 }